Taken from official Gentoo tree to test CUDA_GENERATION variable which is: - not supported by Gentoo in the moment - might not be needed if official OpenCV automated detection of CUDA generation within CMake system works as expected - there is bug raised by me on OpenCV system, but with low priority.
With -DCUDA_GENERATION="Maxwell" I enforced to compile only 50 and 52 architectures and not all 8 architectures available.

This ebuild is not ment to be used as Maxwell architecture is in the moment enforced and is to be removed and merged into Gentoo system later once I agree with OpenCV maintainers about the right way to support this.